requirements.txt
tensorflow-cpu
fastapi
Pillow
starlette
gunicorn
uvicorn
numpy
python-multipart
opencv-python-headless
matplotlib
aiofiles
jinja2